refactor(msgraph): update conditional access policy
