+++ msgraph/conditionalaccesspolicy.go
@@ -1,4 +1,4 @@
-package clients
+package msgraph
 
 import (
 	"context"
@@ -8,32 +8,29 @@ import (
 	"io/ioutil"
 	"net/http"
 	"net/url"
-
-	"github.com/manicminer/hamilton/base"
-	"github.com/manicminer/hamilton/models"
 )
 
 // ConditionalAccessPolicyClient performs operations on ConditionalAccessPolicy.
 type ConditionalAccessPolicyClient struct {
-	BaseClient base.Client
+	BaseClient Client
 }
 
 // NewConditionalAccessPolicyClient returns a new ConditionalAccessPolicyClient
 func NewConditionalAccessPolicyClient(tenantId string) *ConditionalAccessPolicyClient {
 	return &ConditionalAccessPolicyClient{
-		BaseClient: base.NewClient(base.VersionBeta, tenantId),
+		BaseClient: NewClient(VersionBeta, tenantId),
 	}
 }
 
 // List returns a list of ConditionalAccessPolicys, optionally filtered using OData.
-func (c *ConditionalAccessPolicyClient) List(ctx context.Context, filter string) (*[]models.ConditionalAccessPolicy, int, error) {
+func (c *ConditionalAccessPolicyClient) List(ctx context.Context, filter string) (*[]ConditionalAccessPolicy, int, error) {
 	params := url.Values{}
 	if filter != "" {
 		params.Add("$filter", filter)
 	}
-	resp, status, _, err := c.BaseClient.Get(ctx, base.GetHttpRequestInput{
+	resp, status, _, err := c.BaseClient.Get(ctx, GetHttpRequestInput{
 		ValidStatusCodes: []int{http.StatusOK},
-		Uri: base.Uri{
+		Uri: Uri{
 			Entity:      "/identity/conditionalAccess/policies",
 			Params:      params,
 			HasTenantId: true,
@@ -45,7 +42,7 @@ func (c *ConditionalAccessPolicyClient) List(ctx context.Context, filter string)
 	defer resp.Body.Close()
 	respBody, _ := ioutil.ReadAll(resp.Body)
 	var data struct {
-		ConditionalAccessPolicys []models.ConditionalAccessPolicy `json:"value"`
+		ConditionalAccessPolicys []ConditionalAccessPolicy `json:"value"`
 	}
 	if err := json.Unmarshal(respBody, &data); err != nil {
 		return nil, status, err
@@ -54,16 +51,16 @@ func (c *ConditionalAccessPolicyClient) List(ctx context.Context, filter string)
 }
 
 // Create creates a new ConditionalAccessPolicy.
-func (c *ConditionalAccessPolicyClient) Create(ctx context.Context, conditionalAccessPolicy models.ConditionalAccessPolicy) (*models.ConditionalAccessPolicy, int, error) {
+func (c *ConditionalAccessPolicyClient) Create(ctx context.Context, conditionalAccessPolicy ConditionalAccessPolicy) (*ConditionalAccessPolicy, int, error) {
 	var status int
 	body, err := json.Marshal(conditionalAccessPolicy)
 	if err != nil {
 		return nil, status, err
 	}
-	resp, status, _, err := c.BaseClient.Post(ctx, base.PostHttpRequestInput{
+	resp, status, _, err := c.BaseClient.Post(ctx, PostHttpRequestInput{
 		Body:             body,
 		ValidStatusCodes: []int{http.StatusCreated},
-		Uri: base.Uri{
+		Uri: Uri{
 			Entity:      "/identity/conditionalAccess/policies",
 			HasTenantId: true,
 		},
@@ -73,7 +70,7 @@ func (c *ConditionalAccessPolicyClient) Create(ctx context.Context, conditionalA
 	}
 	defer resp.Body.Close()
 	respBody, _ := ioutil.ReadAll(resp.Body)
-	var newConditionalAccessPolicy models.ConditionalAccessPolicy
+	var newConditionalAccessPolicy ConditionalAccessPolicy
 	if err := json.Unmarshal(respBody, &newConditionalAccessPolicy); err != nil {
 		return nil, status, err
 	}
@@ -81,10 +78,10 @@ func (c *ConditionalAccessPolicyClient) Create(ctx context.Context, conditionalA
 }
 
 // Get retrieves an ConditionalAccessPolicy.
-func (c *ConditionalAccessPolicyClient) Get(ctx context.Context, id string) (*models.ConditionalAccessPolicy, int, error) {
-	resp, status, _, err := c.BaseClient.Get(ctx, base.GetHttpRequestInput{
+func (c *ConditionalAccessPolicyClient) Get(ctx context.Context, id string) (*ConditionalAccessPolicy, int, error) {
+	resp, status, _, err := c.BaseClient.Get(ctx, GetHttpRequestInput{
 		ValidStatusCodes: []int{http.StatusOK},
-		Uri: base.Uri{
+		Uri: Uri{
 			Entity:      fmt.Sprintf("/identity/conditionalAccess/policies/%s", id),
 			HasTenantId: true,
 		},
@@ -94,7 +91,7 @@ func (c *ConditionalAccessPolicyClient) Get(ctx context.Context, id string) (*mo
 	}
 	defer resp.Body.Close()
 	respBody, _ := ioutil.ReadAll(resp.Body)
-	var conditionalAccessPolicy models.ConditionalAccessPolicy
+	var conditionalAccessPolicy ConditionalAccessPolicy
 	if err := json.Unmarshal(respBody, &conditionalAccessPolicy); err != nil {
 		return nil, status, err
 	}
@@ -102,7 +99,7 @@ func (c *ConditionalAccessPolicyClient) Get(ctx context.Context, id string) (*mo
 }
 
 // Update amends an existing ConditionalAccessPolicy.
-func (c *ConditionalAccessPolicyClient) Update(ctx context.Context, conditionalAccessPolicy models.ConditionalAccessPolicy) (int, error) {
+func (c *ConditionalAccessPolicyClient) Update(ctx context.Context, conditionalAccessPolicy ConditionalAccessPolicy) (int, error) {
 	var status int
 	if conditionalAccessPolicy.ID == nil {
 		return status, errors.New("cannot update conditionalAccessPolicy with nil ID")
@@ -117,10 +114,10 @@ func (c *ConditionalAccessPolicyClient) Update(ctx context.Context, conditionalA
 	if err != nil {
 		return status, err
 	}
-	_, status, _, err = c.BaseClient.Patch(ctx, base.PatchHttpRequestInput{
+	_, status, _, err = c.BaseClient.Patch(ctx, PatchHttpRequestInput{
 		Body:             body,
 		ValidStatusCodes: []int{http.StatusNoContent},
-		Uri: base.Uri{
+		Uri: Uri{
 			Entity:      fmt.Sprintf("/identity/conditionalAccess/policies/%s", *conditionalAccessPolicy.ID),
 			HasTenantId: true,
 		},
@@ -133,9 +130,9 @@ func (c *ConditionalAccessPolicyClient) Update(ctx context.Context, conditionalA
 
 // Delete removes a ConditionalAccessPolicy.
 func (c *ConditionalAccessPolicyClient) Delete(ctx context.Context, id string) (int, error) {
-	_, status, _, err := c.BaseClient.Delete(ctx, base.DeleteHttpRequestInput{
+	_, status, _, err := c.BaseClient.Delete(ctx, DeleteHttpRequestInput{
 		ValidStatusCodes: []int{http.StatusNoContent},
-		Uri: base.Uri{
+		Uri: Uri{
 			Entity:      fmt.Sprintf("/identity/conditionalAccess/policies/%s", id),
 			HasTenantId: true,
 		},
