+++ msgraph/models.go
@@ -219,6 +219,10 @@ func (a *ApplicationApi) UpdateOAuth2PermissionScope(scope PermissionScope) erro
 	return nil
 }
 
+type ApplicationEnforcedRestrictionsSessionControl struct {
+	IsEnabled *bool `json:"isEnabled,omitempty"`
+}
+
 type ApplicationWeb struct {
 	HomePageUrl           *string                `json:"homePageUrl"`
 	ImplicitGrantSettings *ImplicitGrantSettings `json:"implicitGrantSettings,omitempty"`
@@ -234,6 +238,72 @@ type AppRole struct {
 	IsEnabled          *bool     `json:"isEnabled,omitempty"`
 	Origin             *string   `json:"origin,omitempty"`
 	Value              *string   `json:"value,omitempty"`
+}
+
+type CloudAppSecurityControl struct {
+	IsEnabled            *bool   `json:"isEnabled,omitempty"`
+	CloudAppSecurityType *string `json:"cloudAppSecurityType,omitempty"`
+}
+
+// ConditionalAccessPolicy describes an Conditional Access Policy object.
+type ConditionalAccessPolicy struct {
+	Conditions       *ConditionalAccessConditionSet    `json:"conditions,omitempty"`
+	CreatedDateTime  *time.Time                        `json:"createdDateTime,omitempty"`
+	DisplayName      *string                           `json:"displayName,omitempty"`
+	GrantControls    *ConditionalAccessGrantControls   `json:"grantControls,omitempty"`
+	ID               *string                           `json:"id,omitempty"`
+	ModifiedDateTime *time.Time                        `json:"modifiedDateTime,omitempty"`
+	SessionControls  *ConditionalAccessSessionControls `json:"sessionControls,omitempty"`
+	State            *string                           `json:"state,omitempty"`
+}
+
+type ConditionalAccessConditionSet struct {
+	Applications     *ConditionalAccessApplications `json:"applications,omitempty"`
+	Users            *ConditionalAccessUsers        `json:"users,omitempty"`
+	ClientAppTypes   *[]string                      `json:"clientAppTypes,omitempty"`
+	Locations        *ConditionalAccessLocations    `json:"locations,omitempty"`
+	Platforms        *ConditionalAccessPlatforms    `json:"platforms,omitempty"`
+	SignInRiskLevels *[]string                      `json:"signInRiskLevels,omitempty"`
+	UserRiskLevels   *[]string                      `json:"userRiskLevels,omitempty"`
+}
+
+type ConditionalAccessApplications struct {
+	IncludeApplications *[]string `json:"includeApplications,omitempty"`
+	ExcludeApplications *[]string `json:"excludeApplications,omitempty"`
+	IncludeUserActions  *[]string `json:"includeUserActions,omitempty"`
+}
+
+type ConditionalAccessUsers struct {
+	IncludeUsers  *[]string `json:"includeUsers,omitempty"`
+	ExcludeUsers  *[]string `json:"excludeUsers,omitempty"`
+	IncludeGroups *[]string `json:"includeGroups,omitempty"`
+	ExcludeGroups *[]string `json:"excludeGroups,omitempty"`
+	IncludeRoles  *[]string `json:"includeRoles,omitempty"`
+	ExcludeRoles  *[]string `json:"excludeRoles,omitempty"`
+}
+
+type ConditionalAccessLocations struct {
+	IncludeLocations *[]string `json:"includeLocations,omitempty"`
+	ExcludeLocations *[]string `json:"excludeLocations,omitempty"`
+}
+
+type ConditionalAccessPlatforms struct {
+	IncludePlatforms *[]string `json:"includePlatforms,omitempty"`
+	ExcludePlatforms *[]string `json:"excludePlatforms,omitempty"`
+}
+
+type ConditionalAccessGrantControls struct {
+	Operator                    *string   `json:"operator,omitempty"`
+	BuiltInControls             *[]string `json:"builtInControls,omitempty"`
+	CustomAuthenticationFactors *[]string `json:"customAuthenticationFactors,omitempty"`
+	TermsOfUse                  *[]string `json:"termsOfUse,omitempty"`
+}
+
+type ConditionalAccessSessionControls struct {
+	ApplicationEnforcedRestrictions *ApplicationEnforcedRestrictionsSessionControl `json:"applicationEnforcedRestrictions,omitempty"`
+	CloudAppSecurity                *CloudAppSecurityControl                       `json:"cloudAppSecurity,omitempty"`
+	PersistentBrowser               *PersistentBrowserSessionControl               `json:"persistentBrowser,omitempty"`
+	SignInFrequency                 *SignInFrequencySessionControl                 `json:"signInFrequency,omitempty"`
 }
 
 // Domain describes a Domain object.
@@ -482,6 +552,11 @@ type PermissionScope struct {
 	Value                   *string `json:"value,omitempty"`
 }
 
+type PersistentBrowserSessionControl struct {
+	IsEnabled *bool   `json:"isEnabled,omitempty"`
+	Mode      *string `json:"mode,omitempty"`
+}
+
 type PublicClient struct {
 	RedirectUris *[]string `json:"redirectUris,omitempty"`
 }
@@ -559,6 +634,12 @@ const (
 	SignInAudienceAzureADMultipleOrgs                SignInAudience = "AzureADMultipleOrgs"
 	SignInAudienceAzureADandPersonalMicrosoftAccount SignInAudience = "AzureADandPersonalMicrosoftAccount"
 )
+
+type SignInFrequencySessionControl struct {
+	IsEnabled *bool   `json:"isEnabled,omitempty"`
+	Type      *string `json:"type,omitempty"`
+	Value     *int32  `json:"value,omitempty"`
+}
 
 type SingleSignOnField struct {
 	CustomizedLabel *string `json:"customizedLabel,omitempty"`
